requirements.txt
fastapi
uvicorn
Pillow
pytesseract
python-multipart
sentence-transformers
torch
numpy
